+++ NEWS.md
@@ -81,6 +81,7 @@
 - Preserve ACCESSIBILITY_NOTE (MARC 532) when an Instance is saved [MODLD-728](https://folio-org.atlassian.net/browse/MODLD-728)
 - Allow Import of Authority Records Without LCCN [MODLD-940](https://folio-org.atlassian.net/browse/MODLD-940)
 - Set LINK of the identity resource based on FOLIO configuration [MODLD-923](https://folio-org.atlassian.net/browse/MODLD-923)
+- Include tenant ID in cache [MODLD-952](https://folio-org.atlassian.net/browse/MODLD-952)
 
 ## 1.0.4 (04-24-2025)
 - Work Edit form - Instance read-only section: "Notes about the instance" data is not shown [MODLD-716](https://folio-org.atlassian.net/browse/MODLD-716)